Cells should know their value picker controllers class. Experimenting with cells building the picker delegation and controller value saving.
I'm almost inclined to have the cell be the delegate as well, but it seems risky since cells are intended to be reused and transient. I can't be sure the table view won't destroy/reuse them before the process is finished. But we do know they have all the appropriate state to establish the delegation necessary so greatly simplify the settings view controller.